+++ backend/src/configs/helmet.js
@@ -1,0 +1,14 @@
+const helmet = require('helmet');
+const config = {
+    HELMET_OPTIONS: {
+        contentSecurityPolicy: {
+            useDefaults: true,
+            directives: {
+                "script-src": "'self' 'unsafe-inline'"
+            }
+        }
+    }
+}
+module.exports = (app) => {
+    app.use(helmet(config.HELMET_OPTIONS));
+}