+++ utils/buildForProd.sh
@@ -1,12 +1,33 @@
 clear
+me=`basename "$0"`
 msg="\e[1;32m"
 neutre="\e[0;m"
 DIR="./node_modules"
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
+# Loader
+startLoader() {
+    i=1
+    sp="/-\|"
+    echo -n '  '
+    while true; do
+        printf "\b${sp:i++%${#sp}:1}"
+    done &
+    loader_id=$!
+}
+endLoader() {
+    kill $loader_id
+    printf '\r'
+}
+
+
+echo -e "\n${msg}[${me^^}]${neutre}"
+
 # Suppression du précédent dossier
-echo -e "\n ${msg}suppression du précédent dossier ...${neutre}"
+echo -e "\n ${msg}suppression du précédent dossier ...${neutre} \n"
+startLoader
 rm -rf "${SCRIPT_DIR}/../dist/"
+endLoader
 
 # Build du backend
 echo -e "\n${msg}----- Backend -----${neutre}"
@@ -19,8 +40,10 @@ if [ ! -d "../dist" ]; then
 fi
 cp -r ./!(node_modules) ../dist
 cd ../dist
-echo -e "\n ${msg}installation des dépendances du backend ...${neutre}"
+echo -e "\n ${msg}installation des dépendances du backend ...${neutre}\n"
+startLoader
 npm install
+endLoader
 echo -e "\n ${msg}modification de app.js pour la prise en compte du frontend ...${neutre}"
 sed -i "/catch 404/i\// Angular\napp.use(express.static(path.join(__dirname, \"public\")));\napp.get('**', function (req, res) {\n  res.sendFile(__dirname + '/public/index.html');\n});\n" ../dist/src/app.js
 echo -e "\n ${msg}Fin du build du backend${neutre} \n"
@@ -29,13 +52,17 @@ echo -e "\n ${msg}Fin du build du backend${neutre} \n"
 echo -e "${msg}\n----- Frontend -----${neutre}"
 cd "../frontend"
 if [ ! -d "$DIR" ]; then
-    echo -e "\n ${msg}installation des dépendances du frontend ...${neutre}"
+    echo -e "\n ${msg}installation des dépendances du frontend ...${neutre}\n"
+    startLoader
     npm install
+    endLoader
 else
     echo -e "\n ${msg}dépendances du frontend déjà installées${neutre}"
 fi
 echo -e "\n ${msg}build du frontend ...${neutre}"
+startLoader
 npm run build
+endLoader
 echo -e "\n ${msg}copie des fichiers ...${neutre}"
 mv dist/frontend ../dist/src/public
 echo -e "\n ${msg}Fin du build du frontend${neutre} \n"